+++ common/src/solution.rs
@@ -5,5 +5,5 @@ use crate::*;
 /// Functions that a solution needs to implement.
 pub trait Solution {
     /// Solve the solution with the given `input` and return the result.
-    fn solve(&self, input: &Input) -> u32;
+    fn solve(&self, input: &Input) -> i64;
 }
